How Does An Ultrasonic Sensor Work?
The ultrasonic sensor’s operation is mainly dependent on ultra-sound waves. This is typically the way that Bats are using for vision and navigating around. They send some sound waves in the air, which will reflect back to them to tell how close an object is to the source of the sound wave. You should also know that the ultrasound waves are not detected by the human ears which hear low-frequency tones up to 20kHz.